requirements.txt
japronto
pytest
requests
pyjwt
pika